+++ Sample/Program.cs
@@ -37,7 +37,13 @@ namespace Sample
                 await device.SetTCPParameters(net);
 #endif
                 var basic = await device.GetBasicSettings();
-                Console.WriteLine($"FW {basic.Firmware}");
+                Console.WriteLine($"FW {basic.Firmware} AdminPWD {basic.ManagementPassword} Vol {basic.Volume} DateFormat {basic.DateFormat} 24h {basic.Is24HourClock}");
+#if false //here you can change basic parameters
+                basic.Volume = 0;
+                basic.DateFormat = Anviz.SDK.Responses.DateFormat.DDMMYY;
+                basic.Is24HourClock = true;
+                await device.SetBasicSettings(basic);
+#endif
                 var employees = await device.GetEmployeesData();
                 var dict = new Dictionary<ulong, string>();
                 foreach (var employee in employees)